% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.SDPDm.R
\name{coef.SDPDm}
\alias{coef.SDPDm}
\title{Extract coefficients from model of class SDPDm}
\usage{
\method{coef}{SDPDm}(object, ...)
}
\arguments{
\item{object}{object of class "SDPDm"}

\item{...}{additional arguments to be passed}
}
\value{
Coefficients extracted from the model object of class "SDPDm".
}
\description{
Method for extracting coefficients of objects of class "SDPDm"
}
\seealso{
\code{SDPDm}
}
\author{
Rozeta Simonovska
}
