% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDistMat.R
\name{DDistMat}
\alias{DDistMat}
\title{Double-Power Distance Weights Matrix}
\usage{
DDistMat(distMat, distCutOff = 1e+05, powr = 2)
}
\arguments{
\item{distMat}{distance matrix}

\item{distCutOff}{cut off distance (in metres). Default 100000 metres}

\item{powr}{power (positive exponent), default 2}
}
\value{
\describe{\emph{W}}  weights matrix (not normalised)
}
\description{
This function calculates the double-power distance matrix,
for a given cut off distance and a positive exponent.
}
\details{
W is an \emph{nxn} matrix with elements \eqn{w_{ij}}, \eqn{i, j = 1, ... n}, where
\deqn{w_{ij} = (1-(\frac{d_{ij}}{D})^p)^p, if 0 <= d_{ij} < D}
\deqn{w_{ij} = 0, if d_{ij} > D or i = j}
where D is the cut-off distance point (maximum radius of influence),
\eqn{d_{ij}} is the distance between spatial units \emph{i} and \emph{j},
and \emph{p} is the power value (e.g. \emph{p} = 2, 3, 4,...).
}
\examples{
data(gN3dist)
W1<-DDistMat(distMat=gN3dist)
dist2<-gN3dist/1000 ##in km
W2<-DDistMat(distMat=gN3dist,300000,powr=3) ##distance in metres
W3<-DDistMat(distMat=dist2,300,powr=3)  ##distance in kilometres

}
\author{
Rozeta Simonovska
}
