% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eignor.R
\name{eignor}
\alias{eignor}
\title{Maximum eigen value normalisation}
\usage{
eignor(W)
}
\arguments{
\item{W}{spatial weights matrix}
}
\value{
\describe{\emph{W}}  Eigen normalised spatial weights matrix
}
\description{
Maximum eigen value row normalisation of a spatial weight matrix.
}
\examples{
data(gN3dist)
W<-InvDistMat(distMat=gN3dist,100000,powr=2,metres=FALSE)
Wnor<-eignor(W)

}
\seealso{
\code{\link{rownor}}
}
\author{
Rozeta Simonovska
}
