% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpDistMat.R
\name{ExpDistMat}
\alias{ExpDistMat}
\title{Exponential distance matrix}
\usage{
ExpDistMat(distMat, distCutOff = NULL, expn = 0.01, metres = TRUE)
}
\arguments{
\item{distMat}{distance matrix}

\item{distCutOff}{cut off distance in metres.
If no value is specified, then exponential decay is calulated for all values.}

\item{expn}{positive exponent, default 0.01}

\item{metres}{logical, default TRUE. If FALSE, the distance is transformed into kilometres}
}
\value{
\describe{\emph{W}}  weights matrix (not normalised)
}
\description{
This function calculates the (negative) exponential distance matrix,
with a given cut off distance and a positive exponent value.
}
\details{
W is an \emph{nxn} matrix with elements \eqn{w_{ij}}, \emph{i, j = 1,..n}, where
\deqn{w_{ij}=e^{-\alpha d_{ij}}, if 0 <= d_{ij} < D}
\deqn{w_{ij}=0, if d_{ij} > D or i = j}
where D is the distance cut-off point (maximum radius of influence),
\eqn{d_{ij}} is the distance between spatial units \emph{i} and \emph{j}, and
\eqn{\alpha} is the positive exponent (e.g. \eqn{\alpha}= 0.01, 0.02,...).
}
\examples{
data(gN3dist)
W1<-ExpDistMat(distMat=gN3dist)
W2<-ExpDistMat(distMat=gN3dist,100000,expn=0.001,metres=FALSE)
Wnor<-eignor(W2)

}
\author{
Rozeta Simonovska
}
