% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvDistMat.R
\name{InvDistMat}
\alias{InvDistMat}
\title{Inverse distance matrix}
\usage{
InvDistMat(distMat, distCutOff = NULL, powr = 1, mevn = FALSE)
}
\arguments{
\item{distMat}{distance matrix}

\item{distCutOff}{cut-off distance. Default = half of the maximal distance from the distance matrix.}

\item{powr}{power (positive exponent), default = 1}

\item{mevn}{logical, default FALSE. If TRUE, max-eigenvalue normalisation is performed.}
}
\value{
\describe{\emph{W}}  weights matrix (not normalised)
}
\description{
This function calculates the inverse distances,
with a given cut-off distance and a positive exponent.
}
\details{
W is an \emph{nxn} matrix with elements \eqn{w_{ij}}, \emph{i,j=1,..n}, where
\deqn{w_{ij}=1/d_{ij}^\gamma, if 0 <= d_{ij} < D}
\deqn{w_{ij}=0, if d_{ij} > D or i = j}
where \emph{D} is the distance cut-off point (maximum radius of influence),
\eqn{d_{ij}} is the distance between spatial units \emph{i} and \emph{j},
and \eqn{\gamma} is the value for the exponent (e.g. \eqn{\gamma} = 1, 2, 3, 4,...).
}
\examples{
## distance between centroids of NUTS3 regions in Germany (in metres)
data(gN3dist,package = "SDPDmod")
## inverse distance matrix with cut-off 100000 metres
W1<-InvDistMat(distMat=gN3dist, distCutOff=100000)
dist2<-gN3dist/1000 ##distance in km
## normalised distance matrix with cut-off 100km
W2<-InvDistMat(distMat=dist2, distCutOff=100, powr=2, mevn=TRUE)

}
\author{
Rozeta Simonovska
}
