% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mNearestN.R
\name{mNearestN}
\alias{mNearestN}
\title{m nearest neigbours based on a distance matrix}
\usage{
mNearestN(distMat, m = 5, listv = FALSE, rn = FALSE, zrow = TRUE)
}
\arguments{
\item{distMat}{distance matrix}

\item{m}{number of nearest neigbours, default value 5}

\item{listv}{logocal, default FALSE. If TRUE the list of neighbours should also be returned}

\item{rn}{logical, default FALSE. If TRUE, the weigth matrix will be row normalised}

\item{zrow}{logical, default TRUE. If FALSE, the row matrix is normalised even if there are zero rows in the matrix.}
}
\value{
\describe{\emph{W}} weights matrix
\describe{\emph{nlist}} list of indexes of the m nearest neigbours
}
\description{
This function finds the m nearest neaigbours, given a matrix of distances.
}
\examples{
data(gN3dist,package = "SDPDmod")
fournn<-mNearestN(gN3dist,4)
mat1<-rownor(fournn)
tennn<-mNearestN(gN3dist,10,listv=TRUE,rn=TRUE)
mat2<-tennn$W

}
\author{
Rozeta Simonovska
}
