% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.blmpSDPD.R
\name{print.blmpSDPD}
\alias{print.blmpSDPD}
\title{Print for class blmpSDPD}
\usage{
\method{print}{blmpSDPD}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{object of class "blmpSDPD"}

\item{digits}{number of digits}

\item{...}{additional arguments to be passed}
}
\value{
No return value
}
\description{
Method for printing the results of objects of class "blmpSDPD"
}
\author{
Rozeta Simonovska
}
