% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SDPDm.R
\name{print.SDPDm}
\alias{print.SDPDm}
\title{print for class SDPDm}
\usage{
\method{print}{SDPDm}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{object of class "SDPDm"}

\item{digits}{number of digits}

\item{...}{additional arguments to be passed}
}
\value{
No return value
}
\description{
Method for sprinting the results of objects of class "SDPDm"
}
\seealso{
\code{SDPDm}
}
\author{
Rozeta Simonovska
}
