% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.impactsSDPDm.R
\name{print.summary.impactsSDPDm}
\alias{print.summary.impactsSDPDm}
\title{Print summary for class impactsSDPDm}
\usage{
\method{print}{summary.impactsSDPDm}(x, ...)
}
\arguments{
\item{x}{summary object of class "impactsSDPDm"}

\item{...}{additional arguments to be passed}
}
\description{
Method for printing the summary the results of objects of class "impactsSDPDm"
}
\author{
Rozeta Simonovska
}
