% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.SDPDm.R
\name{print.summary.SDPDm}
\alias{print.summary.SDPDm}
\title{Print of summary for class SDPDm}
\usage{
\method{print}{summary.SDPDm}(x, ...)
}
\arguments{
\item{x}{summary object of class "SDPDm"}

\item{...}{additional arguments to be passed}
}
\value{
No return value
}
\description{
Method for printing the summary the results of objects of class "SDPDm"
}
\seealso{
\code{SDPDm}
}
\author{
Rozeta Simonovska
}
