% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownor.R
\name{rownor}
\alias{rownor}
\title{Row normalisation}
\usage{
rownor(W, zrow = TRUE)
}
\arguments{
\item{W}{spatial weights matrix}

\item{zrow}{logical, default TRUE. If FALSE, the row matrix is normalised even if there are zero rows in the matrix.}
}
\value{
\describe{\emph{W}}  row normalised spatial weights matrix
}
\description{
Row normalisation of a spatial weights matrix.
}
\examples{
library("rgdal")
ger<-readOGR(system.file(dsn="shape",package="SDPDmod"),layer="GermanyNUTS3")
W<-mOrdNbr(ger,3)
Wnor<-rownor(W)

}
\seealso{
\code{\link{eignor}}
}
\author{
Rozeta Simonovska
}
