% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPDm.R
\name{SDPDm}
\alias{SDPDm}
\title{Spaital dynamic panel data lag model with fixed effects maximum likelihood estiomation.}
\usage{
SDPDm(
  formula,
  data,
  W,
  index,
  model,
  effects,
  ldet = NULL,
  lndetspec,
  dynamic = FALSE,
  tlaginfo = list(ind = NULL, tl = FALSE, stl = FALSE),
  LYtrans = FALSE,
  incr = NULL,
  rintrv = TRUE,
  demn = FALSE,
  DIRtrans = FALSE
)
}
\arguments{
\item{formula}{a symbolic description for the (static) model to be estimated, not including the dynamic component}

\item{data}{a data.frame}

\item{W}{spatial weights matrix}

\item{index}{the indexes (names of the varables for the spatil and time component)}

\item{model}{a models to be calculated, c("sar","sdm"), default = "sar"}

\item{effects}{type of fixed effects, c("none","individual","time","twoways"), default ="none"}

\item{ldet}{type of computation of log-determinant, c("full","mc"). Default "full" for smaller problems, "mc" for large problems.}

\item{lndetspec}{specifications for the calucation of the log-determinant for mcmc calucation. Default list(p=30,m=30,sd=12345)}

\item{dynamic}{logical, if TRUE time lag of the dependent variable is included. Default = FALSE}

\item{tlaginfo}{specification for the time lag, default = list(ind=NULL,tl=FALSE,stl=FALSE)
\describe{\emph{ind}} - i-th column in \emph{data} which represents the time lag
\describe{\emph{tl}} - logical, default FALSE. If TRUE \eqn{y_{t-1}} (the lagged dependent variable in time is included)
\describe{\emph{stl}} - logical, default FALSE. If TRUE \eqn{Wy_{t-1}} (the lagged dependent variable in space and time is included)}

\item{LYtrans}{logical, default FALSE. If Lee Yu transformation should be used for bais correction}

\item{incr}{increment for vector of values for rho}

\item{rintrv}{logical, default TRUE, calucates eigenvalues of W. If FALSE, the interval for rho is (-1,1)}

\item{demn}{logical, if Lee-Yu transformation for demeaning of the variables to remove fixed effects is performed (only used in static models). Default FALSE}

\item{DIRtrans}{logical, if direct transformatian of variables should be used. Default, FALSE (only used in dynamic models with "twoways" effects)}
}
\value{
An object of class "SDPDm"
\describe{\emph{coefficients}} coefficients estimate of the model parameters (\emph{coefficients1} for dynamic model)
\describe{\emph{rho}} spatial coefficient
\describe{\emph{sige}} residuals variance
\describe{\emph{llik}} the value of the log likelihood function
\describe{\emph{...}}
}
\description{
This function estimates spatial panel model with fixed effects for static or
dynamic model. It includes the transformation approach suggested by Yu et al (2008) and Lee and Yu (1010).
}
\details{
Based on MatLab functions sar_jihai.m, sar_jihai_time.m and sar_panel_FE.m
}
\examples{
n<-401; t<-15
set.seed(12345)
y<-rnorm(n*t)
x1<-rnorm(n*t)
x2<-rnorm(n*t)
nind<-as.factor(rep(paste0("i",seq(1,n)),t))
tind<-as.factor(rep(paste0("t",seq(1,t)),each=n))
data1<-cbind(nind,tind,y,x1,x2)
data1<-as.data.frame(data1)
fm<-y~x1+x2
ger<-rgdal::readOGR(system.file(dsn="shape",package="SDPDmod"),layer="GermanyNUTS3")
ww<-mOrdNbr(ger)
ww<-rownor(ww)
mod1<-SDPDm(formula=fm, data = data1, W=ww,index=c("nind","tind"),
model="sar",effects = "individual",LYtrans = TRUE)
summary(mod1)
imp<-impactsSDPDm(mod1)

}
\references{
Yu, J., De Jong, R., & Lee, L. F. (2008). Quasi-maximum likelihood estimators for spatial dynamic panel data with fixed effects when both n and T are large. \emph{Journal of Econometrics}, 146(1), 118-134.

Lee, L. F., & Yu, J. (2010). Estimation of spatial autoregressive panel data models with fixed effects. \emph{Journal of Econometrics}, 154(2), 165-185.

Lee, L. F., & Yu, J. (2010). A spatial dynamic panel data model with both time and individual fixed effects. \emph{Econometric Theory}, 564-597.
}
\author{
Rozeta Simonovska
}
