% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedBMat.R
\name{SharedBMat}
\alias{SharedBMat}
\title{Shared boundary matrix}
\usage{
SharedBMat(sf_pol, rn = FALSE)
}
\arguments{
\item{sf_pol}{spatial polygons or spatial lines object}

\item{rn}{logical, default FALSE. If TRUE, the weigth matrix is row normalised}
}
\value{
\describe{\emph{W}} spatial weights matrix (length of shared boundary between spatial units)
}
\description{
This function calculates the shared boundary matrix
}
\examples{
library("rgdal")
ger<-readOGR(system.file(dsn="shape",package="SDPDmod"),layer="GermanyNUTS3")
W<-SharedBMat(ger)

}
\author{
Rozeta Simonovska
}
