% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.impactsSDPDm.R
\name{summary.impactsSDPDm}
\alias{summary.impactsSDPDm}
\title{Summary for class impactsSDPDm}
\usage{
\method{summary}{impactsSDPDm}(object, ...)
}
\arguments{
\item{object}{object of class "impactsSDPDm"}

\item{...}{additional arguments to be passed}
}
\value{
No return value
}
\description{
Method for summarizing the results of objects of class "impactsSDPDm"
}
\seealso{
\code{SDPDm}
}
\author{
Rozeta Simonovska
}
