% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SDPDm.R
\name{summary.SDPDm}
\alias{summary.SDPDm}
\title{Summary for class SDPDm}
\usage{
\method{summary}{SDPDm}(object, ...)
}
\arguments{
\item{object}{object of class "SDPDm"}

\item{...}{additional arguments to be passed}
}
\value{
No return value
}
\description{
Method for summarizing the results of objects of class "SDPDm"
}
\seealso{
\code{SDPDm}
}
\author{
Rozeta Simonovska
}
