% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpDistMat.R
\name{ExpDistMat}
\alias{ExpDistMat}
\title{Exponential distance matrix}
\usage{
ExpDistMat(distMat, distCutOff = NULL, expn = 0.01, mevn = FALSE)
}
\arguments{
\item{distMat}{distance matrix}

\item{distCutOff}{cutoff distance. Default = the maximal value from the distance matrix.}

\item{expn}{positive exponent, default = 0.01}

\item{mevn}{logical, default FALSE. If TRUE, max-eigenvalue normalization is performed.}
}
\value{
\item{W}{spatial weights matrix (Default, not normalized)}
}
\description{
This function calculates the (negative) exponential distance matrix,
with a given cutoff distance and a positive exponent value.
}
\details{
W is an \emph{nxn} matrix with elements \eqn{w_{ij}}, \emph{i, j = 1,..n}, where
\eqn{w_{ij}=e^{-\alpha d_{ij}}}, if \eqn{0 <= d_{ij} < D} and
\eqn{w_{ij}=0}, if \eqn{d_{ij} > D} or \eqn{i = j}.
\emph{D} is the distance cutoff point (maximum radius of influence),
\eqn{d_{ij}} is the distance between spatial units \emph{i} and \emph{j}, and
\eqn{\alpha} is the positive exponent (e.g. \eqn{\alpha}= 0.01, 0.02,...).
}
\examples{
data(gN3dist) ##distance in meters
W1    <- ExpDistMat(distMat = gN3dist, distCutOff = 100000)
dist2 <- gN3dist/1000 ##in km
W2    <- ExpDistMat(distMat = dist2, distCutOff = 100, expn = 0.02)
W2nor <- ExpDistMat(distMat = dist2, 100000, 0.001, mevn = TRUE)

}
\author{
Rozeta Simonovska
}
