% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownor.R
\name{rownor}
\alias{rownor}
\title{Row-normalization}
\usage{
rownor(W)
}
\arguments{
\item{W}{spatial weights matrix}
}
\value{
\item{W}{row-normalized spatial weights matrix}
}
\description{
Row-normalization of a spatial weights matrix.
}
\examples{
library("sf")
ger   <- st_read(system.file(dsn = "shape/GermanyNUTS3.shp",package = "SDPDmod"))
W     <- mOrdNbr(ger, 3)
Wnor  <- rownor(W)

}
\seealso{
\code{\link{eignor}}
}
\author{
Rozeta Simonovska
}
